replaced assert from test_checkers with assertion methods
