+++ test/test_checkers.py
@@ -3,8 +3,10 @@
 import pytest
 import pkg_resources
 
+from test.utils import AssertionTest
 
-class TestCheckers:
+
+class TestCheckers(AssertionTest):
     """Run a series of tests directly against individual checkers.
     This is a companion to the tests in TestScanner."""
 
@@ -39,5 +41,5 @@ class TestCheckers:
                 get_version = checker.load()
 
         result = get_version([""], file_name)
-        assert result["is_or_contains"] == "is"
-        assert result["modulename"] == expected_result
+        self.assertEqual(result["is_or_contains"], "is")
+        self.assertEqual(result["modulename"], expected_result)
