+++ test/test_scanner.py
@@ -14,12 +14,12 @@ import pytest
 
 from cve_bin_tool.cli import Scanner, InvalidFileError
 from cve_bin_tool.cvedb import CVEDB
-from test.utils import download_file
+from test.utils import download_file, AssertionTest
 
 BINARIES_PATH = os.path.join(os.path.abspath(os.path.dirname(__file__)), "binaries")
 
 
-class TestScanner:
+class TestScanner(AssertionTest):
     """Runs a series of tests against our "faked" binaries.
 
     The faked binaries are very small c files containing the same string signatures we use
@@ -51,31 +51,6 @@ class TestScanner:
     @classmethod
     def teardown_class(cls):
         shutil.rmtree(cls.tempdir)
-
-    @staticmethod
-    def assertEqual(actual, expected):
-        if actual != expected:
-            raise AssertionError
-
-    @staticmethod
-    def assertIn(actual, expected):
-        if actual not in expected:
-            raise AssertionError
-
-    @staticmethod
-    def assertNotIn(actual, expected):
-        if actual in expected:
-            raise AssertionError
-
-    @staticmethod
-    def assertIsNone(actual):
-        if actual is not None:
-            raise AssertionError
-
-    @staticmethod
-    def assertIsNotNone(actual):
-        if actual is None:
-            raise AssertionError
 
     def setup_method(self):
         self.cvedb.open()
