+++ test/utils.py
@@ -33,6 +33,38 @@ class TempDirTest(unittest.TestCase):
         shutil.rmtree(cls.tempdir)
 
 
+class AssertionTest:
+    @staticmethod
+    def assertEqual(actual, expected):
+        if actual != expected:
+            raise AssertionError
+
+    @staticmethod
+    def assertNotEqual(actual, expected):
+        if actual == expected:
+            raise AssertionError
+
+    @staticmethod
+    def assertIn(actual, expected):
+        if actual not in expected:
+            raise AssertionError
+
+    @staticmethod
+    def assertNotIn(actual, expected):
+        if actual in expected:
+            raise AssertionError
+
+    @staticmethod
+    def assertIsNone(actual):
+        if actual is not None:
+            raise AssertionError
+
+    @staticmethod
+    def assertIsNotNone(actual):
+        if actual is None:
+            raise AssertionError
+
+
 def download_file(url, target):
     """ helper method to download a file """
     download = urlopen(url)
